requirements.txt
flask
pymongo
python-dotenv
langchain 
langchain-openai
langchain-pinecone